Add retry logic to rhel-reposync.sh
The reposync command has a fixed number of retries for *all* the
packages it will download, at which point it will quit. We don't want to
let it potentially retry indefinitely, but since we're downloading a lot
of packages, it's liable to fail. Therefore, retry the reposync command
itself several times.

Co-authored-by: Neal Gompa <ngompa@datto.com>